Moved frame generation to its own static class.
